+++ packages/core-js/modules/es.string.replace.js
@@ -5,7 +5,7 @@ var toObject = require('../internals/to-object');
 var toLength = require('../internals/to-length');
 var toInteger = require('../internals/to-integer');
 var advanceStringIndex = require('../internals/advance-string-index');
-var regExpExec = require('../internals/regexp-exec');
+var regExpExec = require('../internals/regexp-exec-abstract');
 var max = Math.max;
 var min = Math.min;
 var floor = Math.floor;
