+++ packages/core-js/modules/es.string.split.js
@@ -5,7 +5,7 @@ var anObject = require('../internals/an-object');
 var speciesConstructor = require('../internals/species-constructor');
 var advanceStringIndex = require('../internals/advance-string-index');
 var toLength = require('../internals/to-length');
-var regExpExec = require('../internals/regexp-exec');
+var regExpExec = require('../internals/regexp-exec-abstract');
 var arrayPush = [].push;
 var min = Math.min;
 var LENGTH = 'length';
